revert pohky AgentHUD abi breaking change

+++ FFXIVClientStructs/FFXIV/Client/UI/Agent/AgentHUD.cs
@@ -17,7 +17,7 @@ namespace FFXIVClientStructs.FFXIV.Client.UI.Agent
     public unsafe partial struct AgentHUD
     {
         [FieldOffset(0x0)] public AgentInterface AgentInterface;
-        
+
         //[FieldOffset(0x9C0)] public uint CurrentTargetId;
         //[FieldOffset(0x9C8)] public int TargetCounter;
         //[FieldOffset(0x9D0)] public uint TargetPartyMemberId;
@@ -26,7 +26,7 @@ namespace FFXIVClientStructs.FFXIV.Client.UI.Agent
 
         [FieldOffset(0xBC8)] public int CompanionSummonTimer;
         [FieldOffset(0xBD8)] public fixed byte PartyMemberList[0x20 * 10];
-        [FieldOffset(0xD18)] public short PartyMemberCount;
+        [FieldOffset(0xD18)] public int PartyMemberCount;
         [FieldOffset(0xD20)] public uint PartyTitleAddonId;
         [FieldOffset(0xD24)] public fixed uint RaidMemberIds[40];
         [FieldOffset(0xDC4)] public int RaidGroupSize;
